tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/141.html

<!DOCTYPE html>
<html><head>
  <title>scheduler: SVG inline script that document.writes inline script </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test();
</script>
<svg>
<script>
  t.step(function() {
    log('inline script #1');
    document.write("<" + "script>t.step(function() {log('inline script #2')})<" + "/script>");
    log('end inline script #1');
  });
</script>
</svg>
<script>
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["inline script #1", "inline script #2", "end inline script #1"]);
  t.done();
});
</script>
